Refactor code for api routes
